Refactor repo structure. Add config and env management.
