+++ .env-sample
@@ -1,12 +1,19 @@
-MONGO_URI = ""
-LEPTON_API_KEY = ""
-TOGETHER_API_KEY = ""
-REPLICATE_API_TOKEN = ""
-ANYSCALE_API_KEY = ""
-FIREWORKS_API_KEY = ""
-PERPLEXITY_API_KEY = ""
-DEEPINFRA_API_KEY = ""
-DECART_API_KEY = ""
-OPENAI_API_KEY = ""
-ANTHROPIC_API_KEY = ""
-ABACUS_API_KEY = ""
+ENV=
+
+MONGO_URI=
+AWS_SQS_ARN=
+
+ABACUS_API_KEY=
+ANTHROPIC_API_KEY=
+ANYSCALE_API_KEY=
+DECART_API_KEY=
+DEEPINFRA_API_KEY=
+FIREWORKS_API_KEY=
+LEPTON_API_KEY=
+OPENAI_API_KEY=
+PERPLEXITY_API_KEY=
+REPLICATE_API_TOKEN=
+TOGETHER_API_KEY=
+
+CACHE_EXPIRATION_IN_DAYS=
+HOURS_BETWEEN_COLLECTIONS=
