+++ .env.dev
@@ -1,0 +1,6 @@
+ENV=dev
+
+AWS_SQS_ARN=
+
+CACHE_EXPIRATION_IN_DAYS=0.01
+HOURS_BETWEEN_COLLECTIONS=5
