+++ .env.prod
@@ -1,0 +1,6 @@
+ENV=prod
+
+AWS_SQS_ARN=
+
+CACHE_EXPIRATION_IN_DAYS=0.03
+HOURS_BETWEEN_COLLECTIONS=3
