+++ .env.staging
@@ -1,0 +1,6 @@
+ENV=staging
+
+AWS_SQS_ARN=
+
+CACHE_EXPIRATION_IN_DAYS=0.02
+HOURS_BETWEEN_COLLECTIONS=4
