+++ requirements.txt
@@ -1,14 +1,19 @@
-openai == 1.6.1
-replicate == 0.22.0
-python-dotenv == 1.0.0
-anthropic == 0.3.11
-fastapi == 0.100.0
-async-lru == 2.0.4
-uvicorn == 0.22.0
-motor == 3.2.0
-numpy == 1.25.2
-requests == 2.31.0
-asyncio == 3.4.3
-APScheduler == 3.10.4
-httpx == 0.26.0
-h2 == 4.1.0
+python-decouple==3.8
+python-dotenv==1.0.0
+
+fastapi==0.100.0
+uvicorn==0.29.0
+pydantic==2.7.0
+pydantic-settings==2.2.1
+motor==3.2.0
+requests==2.31.0
+asyncio==3.4.3
+async-lru==2.0.4
+APScheduler==3.10.4
+httpx==0.26.0
+h2==4.1.0
+numpy==1.25.2
+
+openai==1.6.1
+replicate==0.22.0
+anthropic==0.3.11
